refactor: extract requirePlatformAdmin/parseInput to shared helper
Both /platform/tenants/new/actions.ts and /platform/tenants/[id]/actions.ts
defined their own requirePlatformAdmin. Pulled them into one file so the
upcoming editTenantBranding action (in [id]/actions.ts) can reuse the
same admin gate and zod parser as the wizard. No behaviour change.
