+++ apps/web/src/app/platform/tenants/new/actions.ts
@@ -3,31 +3,12 @@ import { db } from "@/db";
 import { tenants, catalogItems } from "@/db/schema";
 import { eq, count } from "drizzle-orm";
 import { revalidatePath } from "next/cache";
-import { getSessionUser, isPlatformAdminEmail } from "@/lib/auth/authorization";
 import { step1Schema, step2Schema, step4Schema } from "@/lib/platform/schema";
 import { getStripe } from "@/lib/stripe";
 import { updateTenantStripe } from "@/db/queries";
 import { cloneCatalogFromTenantUnsafe, type CloneResult } from "@/lib/platform/clone-catalog";
 import { serverCapture } from "@/lib/analytics/server";
-import type { ZodSchema } from "zod";
-
-async function requirePlatformAdmin() {
-  const user = await getSessionUser();
-  if (!user || !isPlatformAdminEmail(user.email)) {
-    throw new Error("Forbidden");
-  }
-  return user;
-}
-
-function parseInput<T>(schema: ZodSchema<T>, input: unknown):
-  | { ok: true; data: T }
-  | { ok: false; error: string } {
-  const r = schema.safeParse(input);
-  if (r.success) return { ok: true, data: r.data };
-  const first = r.error.issues[0];
-  const path = first?.path.join(".");
-  return { ok: false, error: path ? `${path}: ${first.message}` : (first?.message ?? "Invalid input") };
-}
+import { requirePlatformAdmin, parseInput } from "@/lib/platform/action-helpers";
 
 export async function createTenantDraft(input: unknown) {
   const user = await requirePlatformAdmin();
